+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `compute_dwelltimes_using_poly` to `sarkit.cphd`
 - `get_channel_image_area`, `get_extended_image_area`, and `get_scene_image_area` to `sarkit.cphd` and `sarkit.crsd`
 - XML-only SignalNormal check to `sarkit.verification.CphdConsistency.check_channel_signalnormal`
+- GeoInfo/Polygon winding check for SICD
 
 ### Removed
 - Unused `_processing` module
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Erroneous transpose in sensitivity matrix outputs of `sarkit.sicd.projection.compute_gp_xy_parameters`
 - Ignore non-elements in XmlHelper, ElementWrapper, and transcoders
 - Calculation of COA times for bistatic images in `sarkit.sicd.projection.compute_coa_pos_vel`
+- GeoInfo/Polygon winding checks for CRSD and CPHD
 
 
 ## [1.10.1] - 2026-07-31